Refactor requisition detail sections into individual components
