+++ frontend/src/components/requisitionDetail/sections/InfoCardsSection.tsx
@@ -1,0 +1,77 @@
+import React from "react";
+import { Card, List, Skeleton } from "antd";
+import moment from "moment";
+import { Approval } from "../../../types/Approval";
+import { RequisitionSectionProps } from "../RequisitionDetail";
+
+const InfoCardsSection: React.FC<RequisitionSectionProps> = (props) => {
+  const { data, loading } = props;
+  
+  const listData = [
+    {
+      title: "Payment Required By",
+      body: (loading || !data.paymentRequiredBy) ? "Not Set" : moment(data.paymentRequiredBy).format("dddd, MMMM Do, YYYY")
+    },
+    {
+      title: "Created By",
+      body: (loading || !data.createdBy) ? "Not Set" : `${data.createdBy.preferredName} ${data.createdBy.lastName}`
+    },
+    {
+      title: "Vendor",
+      body: (loading || !data.vendor) ? "Not Set" : data.vendor.name
+    }
+  ];
+
+  if (data.approvalSet && data.approvalSet.length > 0) {
+    const approval: Approval = data.approvalSet[data.approvalSet.length - 1]; // Gets last approval
+    let text = "";
+
+    if (approval.isApproving) {
+      text = `Approved by ${approval.approver.preferredName} ${approval.approver.lastName} on ${moment(approval.createdAt).format("M/D/YY")}`;
+    } else {
+      text = `Not approved by ${approval.approver.preferredName} ${approval.approver.lastName} on ${moment(approval.createdAt).format("M/D/YY")} Notes: ${approval.notes}`;
+    }
+
+    listData.push({
+      title: "Approval",
+      body: text
+    });
+  }
+
+  if (data.orderDate) {
+    let text = "";
+
+    if (data.shippingLocation) {
+      text = `Ordered on ${moment(data.orderDate).format("M/D/YY")} and shipped to ${data.shippingLocation}`;
+    } else {
+      text = `Ordered on ${moment(data.orderDate).format("M/D/YY")}`;
+    }
+
+    listData.push({
+      title: "Order Info",
+      body: text
+    });
+  }
+
+  return (
+    <List
+      grid={{ gutter: 16, xs: 1, sm: 3, md: 1, lg: 2, xl: 2, xxl: 2 }}
+      dataSource={listData}
+      id="detail-list"
+      style={{ margin: 0 }}
+      renderItem={(item: any) => (
+        <List.Item>
+          <Card
+            title={item.title}
+            size="small"
+            headStyle={{ wordWrap: "break-word" }}
+          >
+            {loading ? <Skeleton active loading={loading} paragraph={false} /> : item.body}
+          </Card>
+        </List.Item>
+      )}
+    />
+  );
+};
+
+export default InfoCardsSection;
