+++ frontend/src/components/requisitionDetail/sections/expenseRows/ReadyToOrderExpense.tsx
@@ -2,10 +2,10 @@ import React from "react";
 import { useMutation } from "@apollo/client";
 import { Collapse, DatePicker, Form, Input, Tooltip } from "antd";
 import { QuestionCircleOutlined } from "@ant-design/icons/lib";
-import { FORM_RULES } from "../../../util/util";
+import { FORM_RULES } from "../../../../util/util";
 import RequisitionExpenseRow from "./RequisitionExpenseRow";
-import { RequisitionExpenseSectionProps, saveExpenseData } from "../RequisitionExpenseSection";
-import { UPDATE_REQUISITION_MUTATION } from "../../../types/Requisition";
+import { RequisitionExpenseSectionProps, saveExpenseData } from "../ManageStatusSection";
+import { UPDATE_REQUISITION_MUTATION } from "../../../../types/Requisition";
 import CreatePaymentRow from "./CreatePaymentRow";
 
 const ReadyToOrderExpense: React.FC<RequisitionExpenseSectionProps> = (props) => {
