+++ frontend/src/components/requisitionDetail/sections/expenseRows/ReceivedExpense.tsx
@@ -2,8 +2,8 @@ import React from "react";
 import { useMutation } from "@apollo/client";
 import { Collapse } from "antd";
 import RequisitionExpenseRow from "./RequisitionExpenseRow";
-import { RequisitionExpenseSectionProps, saveExpenseData } from "../RequisitionExpenseSection";
-import { UPDATE_REQUISITION_MUTATION } from "../../../types/Requisition";
+import { RequisitionExpenseSectionProps, saveExpenseData } from "../ManageStatusSection";
+import { UPDATE_REQUISITION_MUTATION } from "../../../../types/Requisition";
 import CreatePaymentRow from "./CreatePaymentRow";
 
 const ReceivedExpense: React.FC<RequisitionExpenseSectionProps> = (props) => {
