+++ frontend/src/components/requisitionDetail/sections/expenseRows/RequisitionExpenseRow.tsx
@@ -1,8 +1,8 @@
 import React from "react";
 import { Button, Form, Collapse, Typography } from "antd";
 import { FormInstance } from "antd/es/form";
-import { RequisitionStatus } from "../../../types/Requisition";
-import RequisitionTag from "../../../util/RequisitionTag";
+import { RequisitionStatus } from "../../../../types/Requisition";
+import RequisitionTag from "../../../../util/RequisitionTag";
 
 const { Text } = Typography;
 const { Panel } = Collapse;
